requirements.txt
Django==3.1.5

channels==3.0.3
djangorestframework==3.12.4
dateutils==0.6.6
djangorestframework-simplejwt==4.7.0
mysqlclient==1.4.2

boto3==1.12.15

daphne==3.0.2
gunicorn==20.1.0
